+++ meerschaum/api/dash/callbacks/custom.py
@@ -11,8 +11,13 @@ import traceback
 from meerschaum.api.dash import dash_app
 from meerschaum.plugins import _dash_plugins, _plugin_endpoints_to_pages
 from meerschaum.utils.warnings import warn, dprint
-from meerschaum.api.dash.callbacks.dashboard import _paths, _required_login, _pages
-from meerschaum.api.dash.components import pages_navbar
+from meerschaum.api.dash.callbacks.dashboard import (
+    _paths,
+    _required_login,
+    _pages,
+    _paths_without_dbc_dark,
+)
+from meerschaum.api.dash.components import build_pages_navbar
 
 
 def init_dash_plugins(debug: bool = False):
@@ -42,12 +47,14 @@ def add_plugin_pages(debug: bool = False):
             page_layout = _page_dict['function']()
             if not _page_dict['skip_navbar']:
                 if isinstance(page_layout, list):
-                    page_layout = [pages_navbar] + page_layout
+                    page_layout = [build_pages_navbar()] + page_layout
                 else:
-                    page_layout = [pages_navbar, page_layout]
+                    page_layout = [build_pages_navbar(), page_layout]
             _pages[_page_dict['page_key']] = _endpoint
             if not _endpoint.lstrip('/').startswith('dash'):
                 _endpoint = '/dash/' + _endpoint.lstrip('/')
             _paths[_endpoint] = page_layout
             if _page_dict['login_required']:
                 _required_login.add(_endpoint)
+            if not _page_dict.get('dark_theme', True):
+                _paths_without_dbc_dark.add(_endpoint)