+++ meerschaum/api/dash/pages/tokens.py
@@ -7,13 +7,12 @@ Define the tokens page layout.
 
 import dash_bootstrap_components as dbc
 import dash.html as html
-import dash.dcc as dcc
 from meerschaum._internal.static import STATIC_CONFIG
-from meerschaum.api.dash.components import pages_navbar
+from meerschaum.api.dash.components import build_pages_navbar
 
 
 layout = [
-    pages_navbar,
+    build_pages_navbar(),
     dbc.Container([
         html.Br(), 
         html.H3('Tokens'),
@@ -37,12 +36,14 @@ layout = [
                     color='black',
                     size='sm',
                     id='tokens-refresh-button',
+                    title='Refresh tokens',
                 ),
                 dbc.Button(
                     html.B('+'),
                     color='black',
                     size='sm',
                     id='tokens-create-button',
+                    title='Register token',
                 ),
             ],
             id='tokens-controls-div',